Add virginia age deduction (substraction)
