+++ policyengine_us/parameters/gov/states/va/tax/income/substractions/virginia_age_deduction
@@ -1,0 +1,11 @@
+description: Virginia provides age deduction for people who are 65 and older
+values:
+  2022-01-01: 12_000
+
+metadata:
+  period: year
+  unit: currency-USD
+  label: Virginia age deduction
+  reference:
+  - title: Virginia 2022 Form 760 Resident Individual Income Tax Booklet
+    href: https://www.tax.virginia.gov/sites/default/files/vatax-pdf/2022-760-instructions.pdf#page=16
